refactor modals into individual files
