+++ packages/tokens-studio-for-figma/src/app/components/ManageStylesAndVariables/docsLinks.tsx
@@ -1,0 +1,5 @@
+export const docsLinks = {
+  stylesAndVariables: 'https://docs.tokens.studio/',
+  themes: 'https://docs.tokens.studio/',
+  sets: 'https://docs.tokens.studio/',
+};
